config/requirements/dev.txt
celery
django==4.0.0
django-celery-beat
django-cors-headers
django-redis
djangorestframework
drf-yasg
psycopg2-binary
channels==3.0.4
channels-redis
djangorestframework-simplejwt
djangorestframework-api-key
autopep8
django-debug-toolbar
django-silk
ipdb
ipython
isort
pylint
